changed loss back

+++ tracking_nn.py
@@ -67,7 +67,7 @@ class Net(Module):
         llegx, llegy = (lposh // self.grid).to(self.device), (lposh % self.grid).to(self.device)
         llegh = yh[torch.arange(yh.shape[0]), 4:, llegx, llegy]
 
-        detect_loss = ((rlegh[:, 0] + rlegx - y[:, 0, 0] - y[:, 0, 2]) ** 2 + (rlegh[:, 1] + rlegy - y[:, 0, 1] - y[:, 0, 3]) ** 2 + (llegh[:, 0] + llegx - y[:, 1, 0] - y[:, 1, 2]) ** 2 + (llegh[:, 1] + llegy - y[:, 1, 1] - y[:, 1, 3]) ** 2).sum()
+        detect_loss = ((rlegh[:, 0] + rlegx.double() - y[:, 0, 0] - y[:, 0, 2]) ** 2 + (rlegh[:, 1] + rlegy.double() - y[:, 0, 1] - y[:, 0, 3]) ** 2 + (llegh[:, 0] + llegx.double() - y[:, 1, 0] - y[:, 1, 2]) ** 2 + (llegh[:, 1] + llegy.double() - y[:, 1, 1] - y[:, 1, 3]) ** 2).sum()
 
     def forward(self, x):
         x = x.to(torch.double)
